'copy13'

+++ module-03/forever-young/robotarm-python-2021-main/robotarm-python-2021-main/example - Copy11.py
@@ -4,6 +4,8 @@ robotArm.speed = 3
 for i in range(9):
     robotArm.grab()
     robotArm.scan()
+    if robotArm.scan() == '':
+        break
     if robotArm.scan() == 'white':
         robotArm.moveRight()
         robotArm.drop()
